Change filters to balance speed and discovery
The previous filter found all known test items to run from the explorer
and matched them exactly.
But this won't allow the explorer to discover tests from test results.
If the user selects a group of tests, then it should run all tests in
that group even if they aren't in the explorer
then updated the explorer if there are differences.
This is especially critical for C# tests which can't match code locations
and update that way.