Update release package.json
Update the text color for warnings shown in the status bar when using a light theme. Previously it used #ffcc00, which looks fine with a dark theme, but is very difficult to read when you have a light theme. This pr changes the warning text color for a light theme to #7c6400.